Chess: removed tracker tree DS for tracking solution space
Removed the TrackerTree data structure for tracking solution space that has already been searched

The computation logic to search the solution space has been modified to not traverse already traversed solution space.
This obviates the need to use a data structure such as the TrackerTree to keep track of already traversed sections of the solution space.